+++ axum-extra/src/response/mod.rs
@@ -5,3 +5,7 @@ mod erased_json;
 
 #[cfg(feature = "erased-json")]
 pub use erased_json::ErasedJson;
+
+#[cfg(feature = "ndjson")]
+#[doc(no_inline)]
+pub use crate::ndjson::NdJson;